second

+++ src/main/java/com/example/todolist/repository/ToDoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ToDoRepository extends JpaRepository<ToDo, Long> {
+public interface ToDoRepository extends JpaRepository<ToDo, Integer> {
     ToDo findByID(long id);
 }
